Store dollars instead of cents.
Originally I was storing monetary values as integers in units of cents
to prevent rounding errors.  However, this approach limited how often
budgets could be updated and floating point accuracy is more than enough
for this application.

This commit also simplifies the process of ignoring payments.
Previously a whole boolean column was devoted to marking payments are
either ignored or not.  Now this information is communicated by using
the 'ignore' keyword in the assignment column.  This approach ends up
simplifying both the schema and the code.