Stop casting to int when parsing assignments.
This is a bug from when all balances where integer values in units of
cents.  I since converted to dollars, but forgot to remove this cast.
As a result, the cents from every payment has been ignored up to now.
This commit fixes this bug and adds a test that would have caught it.