Experiment with a bunch of scrapers.
Wells Fargo recently made their website reliant on javascript (and
possibly AJAX) which in turn made it much harder to scrape.  I tried
scraping with WebKit, which understands javascript, but despite a lot of
effort I wasn't able to get that to work.  Then I tried scraping
selenium, which actually runs firefox under the hood, and I was able to
get that to work.

This commit includes the requests, webkit, and selenium scrapers in case
I want to look at them in the future.  However, I will shortly remove
the code I'm not using and just keep the selenium scraper.